requirements.txt
moviepy==1.0.3
SpeechRecognition==3.10.0
openai==1.3.0
python-dotenv==1.0.0
tqdm==4.66.1
yt-dlp